requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.1
django-filter==2.0.0
django-toolbelt==0.0.1
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2==2.7.5
whitenoise==4.1